Add pilight dimmer as light component
