+++ homeassistant/components/pilight/const.py
@@ -1,0 +1,13 @@
+"""Consts used by pilight."""
+from homeassistant.components.light import SUPPORT_BRIGHTNESS
+
+CONF_OFF_CODE = "off_code"
+CONF_OFF_CODE_RECEIVE = "off_code_receive"
+CONF_ON_CODE = "on_code"
+CONF_ON_CODE_RECEIVE = "on_code_receive"
+CONF_SYSTEMCODE = "systemcode"
+CONF_UNIT = "unit"
+CONF_UNITCODE = "unitcode"
+CONF_ECHO = "echo"
+CONF_DIMLEVEL_MIN = 'dimlevel_min'
+CONF_DIMLEVEL_MAX = 'dimlevel_max'
